Make jsmn token struct smaller
Allows to store more tokens in RAM at the cost of bloated .text

+++ partial/libuptiny-demo/common_data.c
@@ -5,7 +5,7 @@
 #define TOKEN_POOL_SIZE 50
 
 jsmntok_t token_pool[TOKEN_POOL_SIZE];
-const unsigned int token_pool_size = TOKEN_POOL_SIZE;
+const int16_t token_pool_size = TOKEN_POOL_SIZE;
 
 #define SIGNATURE_POOL_SIZE 4
 crypto_key_and_signature_t signature_pool[SIGNATURE_POOL_SIZE];
