+++ partial/libuptiny/signatures.h
@@ -8,7 +8,7 @@
 extern "C" {
 #endif
 
-int uptane_parse_signatures(uptane_role_t role, const char *signatures, unsigned int *pos,
+int uptane_parse_signatures(uptane_role_t role, const char *signatures, int16_t *pos,
                             crypto_key_and_signature_t *output, unsigned int max_sigs, uptane_root_t *in_root);
 
 #ifdef __cplusplus
